# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SOS App — Backend API",
    "version": "0.3.0",
    "description": "Snap or Slap challenge app. Sprint 3: Friends & Social System. All endpoints return JSON. No auth — use ?user_id=UUID."
  },
  "servers": [
    { "url": "https://backend-production-2ba1.up.railway.app", "description": "Railway Production" },
    { "url": "http://localhost:3000", "description": "Local Development" }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register new user",
        "description": "Create account with email, password, username. Returns JWT tokens.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "username", "terms_agreed"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "password": { "type": "string", "minLength": 8, "example": "MyPass123!" },
                  "username": { "type": "string", "minLength": 4, "maxLength": 20, "pattern": "^[a-zA-Z0-9_]+$", "example": "cool_user" },
                  "terms_agreed": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created, tokens returned" },
          "400": { "description": "Validation error" },
          "409": { "description": "Email or username already taken" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Sign in",
        "description": "Authenticate with email + password. Returns JWT tokens.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "password": { "type": "string", "example": "MyPass123!" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, tokens returned" },
          "400": { "description": "Validation error" },
          "401": { "description": "Invalid credentials" },
          "403": { "description": "Account disabled" }
        }
      }
    },
...
    "/api/friends": {
      "get": {
        "tags": ["Friends & Social"],
        "summary": "Get friends list",
        "description": "Paginated list of accepted friends. Test: /api/friends?user_id=UUID",
        "parameters": [
          { "name": "user_id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 50 } }
        ],
        "responses": {
          "200": { "description": "Friends list with pagination", "content": { "application/json": { "schema": { "type": "object", "properties": { "friends": { "type": "array" }, "total": { "type": "integer" }, "page": { "type": "integer" }, "limit": { "type": "integer" } } } } } },
          "400": { "description": "Missing user_id" }
        }
      }
    },
    "/api/friends/request": {
      "post": {
        "tags": ["Friends & Social"],
        "summary": "Send friend request",
        "description": "Send a friend request to another user.",
        "parameters": [
          { "name": "user_id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Sender user ID" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["receiver_id"], "properties": { "receiver_id": { "type": "string", "format": "uuid" } } } } }
        },
        "responses": {
          "201": { "description": "Friend request created" },
          "400": { "description": "Validation error or self-request" },
          "404": { "description": "Receiver not found" },
          "409": { "description": "Already friends or request already sent" }
        }
      }
    },
    "/api/friends/request/{requestId}/respond": {
      "put": {
        "tags": ["Friends & Social"],
        "summary": "Accept or decline friend request",
        "description": "Respond to an incoming friend request.",
        "parameters": [
          { "name": "requestId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "user_id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Receiver user ID" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["action"], "properties": { "action": { "type": "string", "enum": ["accept", "decline"] } } } } }
        },
        "responses": {
          "200": { "description": "Request accepted/declined" },
          "400": { "description": "Invalid action" },
          "403": { "description": "Not the receiver" },
          "404": { "description": "Request not found" },
          "409": { "description": "Already processed" }
        }
      }
    },
    "/api/friends/requests/pending": {
      "get": {
        "tags": ["Friends & Social"],
        "summary": "Get pending friend requests",
        "description": "Incoming pending requests. Test: /api/friends/requests/pending?user_id=UUID",
        "parameters": [
          { "name": "user_id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Pending requests with sender info", "content": { "application/json": { "schema": { "type": "object", "properties": { "requests": { "type": "array", "items": { "type": "object", "properties": { "id": { "type": "string" }, "sender": { "type": "object", "properties": { "id": { "type": "string" }, "username": { "type": "string" }, "display_name": { "type": "string" }, "avatar_url": { "type": "string" } } }, "created_at": { "type": "string" } } } } } } } } },
          "400": { "description": "Missing user_id" }
        }
      }
    },
    "/api/friends/{friendUserId}": {
      "delete": {
        "tags": ["Friends & Social"],
        "summary": "Unfriend a user",
        "description": "Remove an existing friendship.",
        "parameters": [
          { "name": "friendUserId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "user_id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Unfriended successfully" },
          "400": { "description": "Missing user_id" },
          "404": { "description": "Not friends" }
        }
      }
    },
...
  }
}